+++ src/documentation/release_notes.md
@@ -24,6 +24,8 @@ The following has been changed or fixed since Jaybird 4.0.3
 -   Fixed: Logic error could lead to incorrect logging of _"Specified statement
     was not created by this connection"_ ([jaybird#674](https://github.com/FirebirdSQL/jaybird/issues/674))
 -   Changed: Updated Firebird 4.0 reserved words based on 4.0.0.2496 ([jaybird#597](https://github.com/FirebirdSQL/jaybird/issues/597))
+-   Fixed: Protocol 15 and 16 had same priority, so Firebird 4.0 might select 
+    protocol 15, leading to timeout support not available. ([jaybird#676](https://github.com/FirebirdSQL/jaybird/issues/676))
 
 Jaybird 4.0.3
 -------------